requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
defusedxml==0.7.1
diff-match-patch==20200713
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-filter==21.1
django-import-export==2.6.1
django-pandas==0.6.6
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
importlib-metadata==4.11.2
Markdown==3.3.6
MarkupPy==1.14
numpy==1.22.2
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.1
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tablib==3.0.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
wincertstore==0.2
xlwt==1.3.0
zipp==3.7.0
